% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_last.R
\name{copy_last}
\alias{copy_last}
\title{Copy path to latest file}
\usage{
copy_last(path = "~/Downloads", glob = NULL)
}
\arguments{
\item{path}{A character vector of one or more paths to find files in.}

\item{glob}{A wildcard aka globbing pattern (e.g. *.csv) passed on to
\code{\link[=grep]{grep()}} to filter paths.}
}
\value{
Returns a path as a character string and copies it to the clipboard.
}
\description{
Copies the path to the newest file in a folder to the clipboard.
}
