% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_shrtcts.R
\name{edit_shrtcts}
\alias{edit_shrtcts}
\title{Edit shrtcts configuration file}
\usage{
edit_shrtcts()
}
\value{
Path to the file, invisibly
}
\description{
Edit shrtcts configuration file
}
\seealso{
\code{\link[shrtcts:add_rstudio_shortcuts]{shrtcts::add_rstudio_shortcuts()}}
}
