% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghosts.R
\name{ghosts}
\alias{ghosts}
\title{Print ghosts in the console}
\usage{
ghosts()
}
\value{
Invisibly returns the string "ghosts"
}
\description{
To run on startup, add the following to your \verb{~/.Rprofile}
(see \code{\link[usethis:edit]{usethis::edit_r_profile()}}):\preformatted{if (interactive() && requireNamespace("arhfn", quietly = TRUE)) \{
  setHook("rstudio.sessionInit", function(isNewSession) \{
    arhfn::ghosts()
  \}, action = "append")
\}
}
}
\examples{
ghosts()
}
\seealso{
\code{\link[=random_art]{random_art()}} to print random color art
}
