% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbl_to_chr.R
\name{lbl_to_chr}
\alias{lbl_to_chr}
\title{lbl_to_chr}
\usage{
lbl_to_chr(tbl)
}
\arguments{
\item{tbl}{Tibble to modify}
}
\value{
A tibble with all labelled columns replaced with character columns
}
\description{
Convert labelled tibble columns to characters
}
