% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_art.R
\name{random_art}
\alias{random_art}
\title{Print random color artwork in the console}
\usage{
random_art()
}
\value{
Invisibly returns the function name of the artwork
}
\description{
Randomly runs either \code{\link[=ghosts]{ghosts()}} or \code{\link[=invaders]{invaders()}}.
}
\details{
To run on startup, add the following to your \verb{~/.Rprofile}
(see \code{\link[usethis:edit]{usethis::edit_r_profile()}}):\preformatted{if (interactive() && requireNamespace("arhfn", quietly = TRUE)) \{
  setHook("rstudio.sessionInit", function(isNewSession) \{
    arhfn::random_art()
  \}, action = "append")
\}
}
}
\examples{
random_art()
}
