% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rda.R
\name{read_rda}
\alias{read_rda}
\title{read_rda}
\usage{
read_rda(file)
}
\arguments{
\item{file}{RData file to read in}
}
\value{
A tibble comtained in the Rdata file
}
\description{
Read rda file
}
\details{
Imports an rda file \emph{à la} the imports functions in \strong{readr} and \strong{haven}
}
