% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgrade}
\alias{upgrade}
\title{Upgrade all installed packages}
\usage{
upgrade(job = TRUE, daily = FALSE)
}
\arguments{
\item{job}{If \code{TRUE}, run the upgrade in a background RStudio job}

\item{daily}{If \code{TRUE}, exit early if \code{upgrade()} has already been run today}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Upgrade all installed packages
}
