% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_arh_drake.R
\name{use_arh_drake}
\alias{use_arh_drake}
\title{use_arh_drake}
\usage{
use_arh_drake()
}
\value{
Nothing. Modifies your workspace.
}
\description{
Setup a drake project
}
\details{
Creates files and directories according to the arh_drake template.
}
