% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_calendar_version.R
\name{use_calendar_version}
\alias{use_calendar_version}
\alias{use_calver}
\title{Increment package version using CalVer}
\usage{
use_calendar_version(which = NULL)

use_calver(which = NULL)
}
\arguments{
\item{which}{If \code{which} is \code{"dev"}, \code{.9000} is appended to the end of the
version number}
}
\value{
Updates the version number in the \code{DESCRIPTION} and \code{NEWS} file.
}
\description{
A complement to \code{usethis::use_version()} that uses
\href{https://calver.org/}{Calendar Versioning}.
}
\seealso{
\code{\link[usethis:use_version]{usethis::use_version()}}
}
