% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dependency.R
\name{use_dependency}
\alias{use_dependency}
\title{Add a package dependency}
\usage{
use_dependency(package, min_version = TRUE)
}
\arguments{
\item{package}{Unquoted or quoted name of package to add as dependency}

\item{min_version}{Optionally, supply a minimum version for the package.
Defaults to \code{TRUE}, which uses the currently installed version.
If \code{FALSE} or \code{NULL}, no minimum is set.}
}
\value{
Adds \code{package} to file \code{DESCRIPTION} as a dependency with the minimum
version set to the currently installed version
}
\description{
Add a package dependency
}
