% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_description.R
\name{use_description}
\alias{use_description}
\title{Open or create a DESCRIPTION file}
\usage{
use_description(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[usethis:use_description]{usethis::use_description()}}}
}
\description{
If a \code{DESCRIPTION} file already exists, opens it.
If not, creates a \code{DESCRIPTION} file with \code{\link[usethis:use_description]{usethis::use_description()}}
}
