% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_example.R
\name{use_example}
\alias{use_example}
\title{Create or edit an example .R file}
\usage{
use_example(name = NULL, dir = "man/examples")
}
\arguments{
\item{name}{File name, without extension.
If \code{\link{NULL}}, the default, automatically generate a name based on the file
that is currently open.}
}
\description{
Creates an R script in \code{man/examples} to store examples for a function.
}
\seealso{
\code{\link[usethis:use_r]{usethis::use_r()}} and \code{\link[usethis:use_r]{usethis::use_test()}}
}
