% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_fn.R
\name{use_fn}
\alias{use_fn}
\title{Create or edit a .R file}
\usage{
use_fn(name, ..., folder = "R")
}
\arguments{
\item{name}{Function name, either as a string or an unquoted name}

\item{...}{Unquoted function arguments}

\item{folder}{The folder to place the file in. Defaults to "R".}
}
\description{
Create or edit a .R file
}
\examples{

\dontrun{
use_fn(test_function, x, y = "z")
}
}
\seealso{
\code{\link[usethis:use_r]{usethis::use_r()}}
}
