% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_mit_badge.R
\name{use_mit_badge}
\alias{use_mit_badge}
\title{README badge for MIT license}
\usage{
use_mit_badge()
}
\description{
README badge for MIT license
}
\examples{

\dontrun{
use_mit_badge()
}
}
\seealso{
\code{\link[usethis:badges]{usethis::use_badge()}}
}
