% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_package.R
\name{use_package}
\alias{use_package}
\alias{use_depends}
\alias{use_suggests}
\title{Add a package dependency}
\usage{
use_package(..., type = "Imports", min_version = NULL, tidy = TRUE)

use_depends(..., min_version = TRUE, tidy = TRUE)

use_suggests(..., min_version = NULL, tidy = TRUE)
}
\arguments{
\item{type}{Whether to add the package as \code{"Imports"}, \code{"Depends"},
or \code{"Suggests"} (case insensitive).}

\item{min_version}{Optionally, supply a minimum version for the package.
If \code{TRUE}, uses the currently installed version.
If \code{FALSE} or \code{NULL}, no minimum is set.
Defaults to \code{NULL} for \code{use_package()} and \code{use_suggests()} and
\code{TRUE} for \code{use_depends()}.}

\item{tidy}{Whether to run \code{\link[usethis:tidyverse]{usethis::use_tidy_description()}} after adding
dependency. Defaults to \code{TRUE}.}

\item{package}{Unquoted or quoted name of package.}
}
\value{
Adds \code{package} to file \code{DESCRIPTION}
}
\description{
Add a package dependency
}
