% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_example.R
\name{use_r}
\alias{use_r}
\title{Create or edit an R file}
\usage{
use_r(name = NULL, open = rlang::is_interactive())
}
\arguments{
\item{name}{Either a name without extension, or \code{NULL} to create a test file
file based on currently open file in the script editor.}

\item{open}{Whether to open the file for interactive editing.}
}
\value{
If an R file does not already exist, creates an R file.
If \code{open} is \code{TRUE}, opens the R file.
}
\description{
Create or edit an R file
}
\seealso{
\code{\link[usethis:use_r]{usethis::use_r()}}, which this function is modified from
}
