% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_example.R
\name{use_test}
\alias{use_test}
\title{Create or edit a test file}
\usage{
use_test(name = NULL, open = rlang::is_interactive())
}
\arguments{
\item{name}{Either a name without extension, or \code{NULL} to create a test file
file based on currently open file in the script editor.}

\item{open}{Whether to open the file for interactive editing.}
}
\value{
If a test file does not already exist, creates a test file.
If \code{open} is \code{TRUE}, opens the test file.
}
\description{
Create or edit a test file
}
\seealso{
\code{\link[usethis:use_r]{usethis::use_test()}}, which this function is modified from
}
