% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_theme_a11y.R
\name{use_theme_a11y}
\alias{use_theme_a11y}
\title{Apply an accessible RStudio theme}
\usage{
use_theme_a11y(dark = NULL, quietly = FALSE)
}
\arguments{
\item{dark}{A logical indicating whether to use a dark theme.
If \code{NULL}, the default, the decision is based on whether the current active
theme is a light theme or a dark theme.}

\item{quietly}{Suppress confirmation messages}
}
\description{
Applies the \link[rsthemes:rsthemes-package]{rsthemes} a11y-light and a11y-dark.
}
