% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_example.R
\name{use_example}
\alias{use_example}
\title{Create or edit an example .R file}
\usage{
use_example(name = NULL, dir = "man/examples", open = rlang::is_interactive())
}
\arguments{
\item{name}{Either a name without extension, or \code{NULL} to create the paired
file based on currently open file in the script editor.
If the \verb{R/} or test file is open, \code{use_example()} will create/open the
corresponding example file.}

\item{dir}{The directory where example scripts will be stored.
Defaults to \code{man/examples}.}

\item{open}{Whether to open the file for interactive editing.}
}
\description{
Creates an R script in \code{man/examples} to store examples for a function.
}
\seealso{
\code{\link[usethis:use_r]{usethis::use_r()}} and \code{\link[usethis:use_r]{usethis::use_test()}}
}
