% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_example.R
\name{use_example}
\alias{use_example}
\title{Create or edit an example .R file}
\usage{
use_example(name = NULL)
}
\arguments{
\item{name}{File name, without extension; will create if it doesn't already
exist. If not specified, and you're currently in a .R file, will guess
name based on .R name.}
}
\description{
Create or edit an example .R file
}
\seealso{
\code{\link[usethis:use_r]{usethis::use_r()}} and \code{\link[usethis:use_test]{usethis::use_test()}}
}
