% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dependency.R
\name{use_package}
\alias{use_package}
\alias{use_dependency}
\title{Add a package dependency}
\usage{
use_package(
  package,
  type = c("Imports", "Depends", "Suggests"),
  min_version = NULL,
  tidy = TRUE,
  character_only = FALSE
)

use_dependency(
  package,
  min_version = TRUE,
  tidy = TRUE,
  character_only = FALSE
)
}
\arguments{
\item{package}{Unquoted or quoted name of package.}

\item{type}{Whether to add the package as an import, dependency,
or suggestion (case insensitive).}

\item{min_version}{Optionally, supply a minimum version for the package.
If \code{TRUE}, uses the currently installed version.
If \code{FALSE} or \code{NULL}, no minimum is set.
Defaults to \code{NULL} for \code{use_package()} and \code{TRUE} for \code{use_dependency()}.}

\item{tidy}{Whether to run \code{\link[usethis:tidyverse]{usethis::use_tidy_description()}} after adding
dependency. Defaults to \code{TRUE}.}

\item{character_only}{A logical indicating whether \code{package} can be assumed
to be a character string.}
}
\value{
Adds \code{package} to file \code{DESCRIPTION}
}
\description{
Add a package dependency
}
